Fix snippet_validator import in models.py
