+++ app/blog/validators.py
@@ -1,0 +1,15 @@
+import re
+from django.core.exceptions import ValidationError
+
+link_media_pattern = re.compile(
+    r"<a.*?/a>|<img.*?/img>|<video.*?/video>|<audio.*?/audio>", flags=re.IGNORECASE
+)
+max_length = 400
+
+
+def snippet_validator(value):
+    value_without_links_media = link_media_pattern.sub("", value)
+    if len(value_without_links_media) > max_length:
+        raise ValidationError(
+            f"The snippet cannot have more than {max_length} characters (excluding links and media)."
+        )
